pyproject.toml
[build-system]
requires = ["poetry==0.12.12"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "test_q"
description = "test_q"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.9.0"

Django = "^3.1.4"
django-split-settings = "^1.0.1"
django-extensions = "^3.1.0"
djangorestframework = "^3.12.2"
django-rest-auth = "^0.9.5"
django-cors-headers = "^3.5.0"
django-redis = "^4.12.1"
django-autoslug = "^1.9.8"

psycopg2-binary = "^2.8.6"
python-decouple = "^3.3"
bcrypt = "^3.2.0"
structlog = "^20.1.0"
tqdm = "^4.54.1"
ujson = "^4.0.1"
requests = "^2.25.0"
pytils = "^0.3"
lxml = "^4.6.2"



[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
